uadk: support sgl memory for nosva mode
When use sgl memory with uadk, we needs to convert
the user SGL into a hardware SGL.
For performance considerations, we creates an SGL
pool for each queue, from which SGL are requested
when performing tasks.
In the nosva scenario, to obtain the mapping the
relationship between virtual and physical addresses,
SGL memory needs to be allocated from the memory pool.

Signed-off-by: Wenkai Lin <linwenkai6@hisilicon.com>